pyproject.toml
[tool.poetry]
name = "project-three"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
mypy = "^1.18.2"
black = "^25.11.0"
flake8 = "^7.3.0"
psycopg2 = "^2.9.11"
django-countries = "^8.2.0"
redis = "^7.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
